Refactor container demos for charts (charts-537) (#415)
Removed VSeven from demos
Created section in demo for Vaadin 7 Container demos
Updated names for DataProvider demos
Updated references for affected tests
