+++ examples/src/main/java/com/vaadin/addon/charts/examples/container/ContainerSeriesWithHighAndLow.java
@@ -1,7 +1,21 @@
 package com.vaadin.addon.charts.examples.container;
 
-import java.util.ArrayList;
-import java.util.Collection;
+/*
+ * #%L
+ * Vaadin Charts
+ * %%
+ * Copyright (C) 2012 - 2015 Vaadin Ltd
+ * %%
+ * This program is available under Commercial Vaadin Add-On License 3.0
+ * (CVALv3).
+ *
+ * See the file licensing.txt distributed with this software for more
+ * information about licensing.
+ *
+ * You should have received a copy of the CVALv3 along with this program.
+ * If not, see <https://vaadin.com/license/cval-3>.
+ * #L%
+ */
 
 import com.vaadin.addon.charts.Chart;
 import com.vaadin.addon.charts.examples.AbstractVaadinChartExample;
@@ -9,20 +23,17 @@ import com.vaadin.addon.charts.examples.SkipFromDemo;
 import com.vaadin.addon.charts.model.ChartType;
 import com.vaadin.addon.charts.model.Configuration;
 import com.vaadin.addon.charts.model.DataLabelsRange;
-import com.vaadin.addon.charts.model.DataProviderSeries;
 import com.vaadin.addon.charts.model.PlotOptionsColumnrange;
 import com.vaadin.addon.charts.model.Tooltip;
 import com.vaadin.addon.charts.model.XAxis;
 import com.vaadin.addon.charts.model.YAxis;
-import com.vaadin.data.provider.DataProvider;
-import com.vaadin.data.provider.ListDataProvider;
 import com.vaadin.ui.Component;
+import com.vaadin.v7.addon.charts.model.ContainerDataSeries;
+import com.vaadin.v7.data.Item;
+import com.vaadin.v7.data.util.IndexedContainer;
 
 @SkipFromDemo
 public class ContainerSeriesWithHighAndLow extends AbstractVaadinChartExample {
-
-    private DataProvider<Temperature, ?> temps = new ListDataProvider<>(
-            getMockData());
 
     @Override
     public String getDescription() {
@@ -53,53 +64,35 @@ public class ContainerSeriesWithHighAndLow extends AbstractVaadinChartExample {
 
         PlotOptionsColumnrange columnRange = new PlotOptionsColumnrange();
         columnRange.setDataLabels(new DataLabelsRange(true));
-        columnRange.getDataLabels()
-                .setFormatter("function() {return this.y + '°C';}");
+        columnRange.getDataLabels().setFormatter(
+                "function() {return this.y + '°C';}");
         conf.setPlotOptions(columnRange);
 
         conf.getLegend().setEnabled(false);
-        conf.setSeries(createChartDS());
+
+        IndexedContainer vaadinContainer = new IndexedContainer();
+        vaadinContainer.addContainerProperty("min", Number.class, null);
+        vaadinContainer.addContainerProperty("max", Number.class, null);
+        for (Number[] minMaxPair : getRawData()) {
+            Item item = vaadinContainer.getItem(vaadinContainer.addItem());
+            item.getItemProperty("min").setValue(minMaxPair[0]);
+            item.getItemProperty("max").setValue(minMaxPair[1]);
+        }
+
+        ContainerDataSeries data = new ContainerDataSeries(vaadinContainer);
+        data.setLowPropertyId("min");
+        data.setHighPropertyId("max");
+        data.setName("Temperatures");
+
+        conf.setSeries(data);
 
         return chart;
     }
 
-    private DataProviderSeries<Temperature> createChartDS() {
-        DataProviderSeries<Temperature> ds = new DataProviderSeries<>(temps);
-        ds.setLow(Temperature::getMin);
-        ds.setHigh(Temperature::getMax);
-        ds.setName("Temperatures");
-        return ds;
-    }
-
-    private class Temperature {
-        private double min;
-        private double max;
-
-        public Temperature(double min, double max) {
-            this.min = min;
-            this.max = max;
-        }
-
-        public double getMin() {
-            return min;
-        }
-
-        public double getMax() {
-            return max;
-        }
-    }
-
-    private Collection<Temperature> getMockData() {
-
-        Double[][] data = { { -9.7, 9.4 }, { -8.7, 6.5 }, { -3.5, 9.4 },
+    private Number[][] getRawData() {
+        return new Number[][] { { -9.7, 9.4 }, { -8.7, 6.5 }, { -3.5, 9.4 },
                 { -1.4, 19.9 }, { 0.0, 22.6 }, { 2.9, 29.5 }, { 9.2, 30.7 },
                 { 7.3, 26.5 }, { 4.4, 18.0 }, { -3.1, 11.4 }, { -5.2, 10.4 },
                 { -13.5, 9.8 } };
-
-        Collection<Temperature> temps = new ArrayList<>();
-        for (int i = 0; i < data.length; i++) {
-            temps.add(new Temperature(data[i][0], data[i][1]));
-        }
-        return temps;
     }
 }
