--- integration-tests/src/test/java/com/vaadin/addon/charts/testbenchtests/ContainerSeriesWithSplineTBTest.java
@@ -1,18 +1,0 @@
-package com.vaadin.addon.charts.testbenchtests;
-
-import com.vaadin.addon.charts.examples.container.ContainerSeriesWithSpline;
-
-public class ContainerSeriesWithSplineTBTest extends
-        AbstractSimpleScreenShotTestBenchTest {
-
-    @Override
-    protected String getTestViewName() {
-        return ContainerSeriesWithSpline.class.getSimpleName();
-    }
-
-    @Override
-    protected String getPackageName() {
-        return "container";
-    }
-
-}
