--- integration-tests/src/test/java/com/vaadin/addon/charts/testbenchtests/VsevenChartWithExternalContainerTBTest.java
@@ -1,17 +1,0 @@
-package com.vaadin.addon.charts.testbenchtests;
-
-import com.vaadin.addon.charts.examples.container.VsevenChartWithExternalContainer;
-
-public class VsevenChartWithExternalContainerTBTest extends
-        AbstractSimpleScreenShotTestBenchTest {
-
-    @Override
-    protected String getTestViewName() {
-        return VsevenChartWithExternalContainer.class.getSimpleName();
-    }
-
-    @Override
-    protected String getPackageName() {
-        return "container";
-    }
-}
