+++ integration-tests/src/test/resources/screenshots/reference/DataProviderSeriesWithSpline_linux_phantomjs_2.png
