+++ integration-tests/src/test/resources/screenshots/reference/DataProviderSeriesWithSpline_windows_firefox_45.png
