+++ integration-tests/src/test/resources/screenshots/reference/DataProviderSeriesWithSpline_windows_internet explorer_11.png
