+++ integration-tests/src/test/resources/screenshots/reference/DataProviderSeriesWithSpline_xp_chrome_54.png
